feat: show the winner in the screen
