+++ src/components/garage/button/button.css
@@ -1,0 +1,40 @@
+.button-wrapper {
+    display: flex;
+    align-items: center;
+    margin-top: 10px;
+}
+
+.reset-button {
+    margin: 0 5px;
+}
+
+.race-button,
+.reset-button,
+.generate-button {
+    font-size: 15px;
+    font-weight: bold;
+    background: lavender;
+    height: 27px;
+    width: 85px;
+    border-radius: 5px;
+    border-bottom: 4px solid cornflowerblue;
+}
+
+.race-button,
+.reset-button {
+    height: 30px;
+    color: black;
+    background: palegoldenrod;
+    border-bottom: 4px solid yellowgreen;
+}
+
+.generate-button {
+    height: 30px;
+    width: 170px;
+}
+
+.reset-button:active,
+.race-button:active,
+.generate-button:active {
+    border-bottom: 2px solid cornflowerblue;
+}
