+++ src/components/garage/car/amount/amount.css
@@ -1,0 +1,4 @@
+.cars-amount {
+    margin-top: 30px;
+    font-size: 30px;
+}
