+++ src/components/garage/car/button/button.css
@@ -1,0 +1,49 @@
+.select-button:active,
+.remove-button:active {
+    border-bottom: 2px solid cornflowerblue;
+}
+
+.a-button,
+.b-button {
+    font-size: 20px;
+    background: transparent;
+    border: 2px solid darkseagreen;
+    border-radius: 5px;
+    font-weight: bold;
+}
+
+.a-button {
+    color: darkseagreen;
+    border: 1px solid darkseagreen;
+}
+
+.b-button {
+    color: saddlebrown;
+    border: 1px solid saddlebrown;
+}
+
+.remove-button,
+.b-button {
+    margin: 0 5px;
+}
+
+.select-button,
+.remove-button {
+    font-size: 15px;
+    font-weight: bold;
+    background: lavender;
+    height: 27px;
+    width: 85px;
+    border-radius: 5px;
+    border-bottom: 4px solid cornflowerblue;
+}
+
+.select-button,
+.remove-button {
+    height: 30px;
+}
+
+.gray {
+    color: darkslategray;
+    border: 1px solid darkslategray;
+}
