+++ src/components/garage/car/car.css
@@ -1,0 +1,25 @@
+@import "flag/flag.css";
+@import "amount/amount.css";
+@import "button/button.css";
+
+.car-wrapper {
+    margin: 20px 0;
+}
+
+.car-wrapper__header {
+    display: flex;
+    align-items: center;
+    margin-bottom: 10px;
+}
+
+.car-name {
+    font-size: 20px;
+    color: darkkhaki;
+    margin-left: 20px;
+}
+
+.car-wrapper__body {
+    display: flex;
+    align-items: center;
+    border-bottom: 1px dashed white;
+}
