+++ src/components/garage/car/flag/flag.css
@@ -1,0 +1,3 @@
+.flag-wrapper {
+    margin: 0 150px 0 auto;
+}
