+++ src/components/garage/winner/winner.ts
@@ -1,0 +1,5 @@
+import './winner.css'
+
+const winner = `<div class="winner display" id="winner"></div>`;
+
+export default winner;
